Initial Rework of Jammer Code
Have made some changes to the Jammer code and the dot.

The dot now goes to the location of the decoyCoordinate.

The Jammer has a different method for generating numbers, unsure if this is closer to Rayleigh or not. The Jammer can now generate Decoy Coordinates within the radius of the jammer. And excess pseudocode has been trimmed.